fix type mismatch between css props in Theme UI components dependent on whether sx in scope or not
